requirements.txt
dj-database-url==0.5.0
Django==3.1.14
django-cors-headers==3.2.0
django-picklefield==2.0
django-q==1.3.4
django-request-logging==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.11.2
django-cors-headers==3.2.0
gunicorn==20.0.4
requests==2.25.0
pandas
django-heroku
whitenoise